+++ errors.go
@@ -353,7 +353,7 @@ func NewStaleObjectError(kind string, id string, actualVersion, expectedVersion 
 	}
 }
 
-func (s *Service) foundationErrorToStatusInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
+func (s *Service) foundationErrorToStatusUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
 	h, err := handler(ctx, req)
 
 	if err != nil {
