+++ grpc/request_logger.go
@@ -9,9 +9,9 @@ import (
 	fctx "github.com/ri-nat/foundation/context"
 )
 
-// LoggingInterceptor returns a gRPC unary interceptor that logs all incoming gRPC calls.
+// LoggingUnaryInterceptor returns a gRPC unary interceptor that logs all incoming gRPC calls.
 // It logs the method details, request, response, and any potential errors.
-func LoggingInterceptor(log *logrus.Entry) func(context.Context, interface{}, *grpc.UnaryServerInfo, grpc.UnaryHandler) (interface{}, error) {
+func LoggingUnaryInterceptor(log *logrus.Entry) func(context.Context, interface{}, *grpc.UnaryServerInfo, grpc.UnaryHandler) (interface{}, error) {
 	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp interface{}, err error) {
 		// Enhance the log with request-related fields.
 		log = log.WithFields(logrus.Fields{
